Start refactoring things into separate modules
